Move worker daemon expiration to WorkerDaemonClientsManager
