+++ subprojects/workers/src/test/groovy/org/gradle/workers/internal/WorkerDaemonClientsManagerTest.groovy
@@ -24,6 +24,7 @@ import org.gradle.internal.event.ListenerManager
 import org.gradle.internal.logging.LoggingManagerInternal
 import org.gradle.internal.logging.events.LogLevelChangeEvent
 import org.gradle.internal.logging.events.OutputEventListener
+import org.gradle.process.internal.health.memory.MemoryManager
 import org.gradle.util.ConcurrentSpecification
 import spock.lang.Subject
 
@@ -36,8 +37,9 @@ class WorkerDaemonClientsManagerTest extends ConcurrentSpecification {
     def serverImpl = Stub(WorkerProtocol)
     def listenerManager = Stub(ListenerManager)
     def loggingManager = Stub(LoggingManagerInternal)
+    def memoryManager = Mock(MemoryManager)
 
-    @Subject manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager)
+    @Subject manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager, memoryManager)
 
     def "does not reserve idle client when no clients"() {
         expect:
@@ -91,7 +93,7 @@ class WorkerDaemonClientsManagerTest extends ConcurrentSpecification {
 
     def "can stop session-scoped clients"() {
         listenerManager = new DefaultListenerManager()
-        manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager)
+        manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager, memoryManager)
         def client1 = Mock(WorkerDaemonClient)
         def client2 = Mock(WorkerDaemonClient)
         starter.startDaemon(serverImpl.class, options) >>> [client1, client2]
@@ -110,7 +112,7 @@ class WorkerDaemonClientsManagerTest extends ConcurrentSpecification {
 
     def "Stopping session-scoped clients does not stop other clients"() {
         listenerManager = new DefaultListenerManager()
-        manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager)
+        manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager, memoryManager)
         def client1 = Mock(WorkerDaemonClient)
         def client2 = Mock(WorkerDaemonClient)
         starter.startDaemon(serverImpl.class, options) >>> [client1, client2]
@@ -158,7 +160,7 @@ class WorkerDaemonClientsManagerTest extends ConcurrentSpecification {
         loggingManager.getLevel() >> LogLevel.INFO
 
         when:
-        manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager)
+        manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager, memoryManager)
 
         then:
         listener != null
@@ -229,5 +231,21 @@ class WorkerDaemonClientsManagerTest extends ConcurrentSpecification {
         0 * client1.stop()
         0 * client2.stop()
         1 * client3.stop()
+    }
+
+    def "registers/deregisters a worker daemon expiration with the memory manager"() {
+        WorkerDaemonExpiration workerDaemonExpiration
+
+        when:
+        def manager = new WorkerDaemonClientsManager(starter, listenerManager, loggingManager, memoryManager)
+
+        then:
+        1 * memoryManager.addMemoryHolder(_) >> { args -> workerDaemonExpiration = args[0] }
+
+        when:
+        manager.stop()
+
+        then:
+        1 * memoryManager.removeMemoryHolder(_) >> { args -> assert args[0] == workerDaemonExpiration }
     }
 }
